# src/main/resources/swagger/minh-shop.yml
openapi: '3.0.3'
info:
  title: 'Minh Shop'
  version: 0.0.1
  description: 'This is a Minh Shop server.  You can find out more about     Swagger
      at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For
      this sample, you can use the api key `special-key` to test the authorization     filters.'
  termsOfService: http://swagger.io/terms/
  contact:
    name: yuvytung
    email: yuvytung@gmail.com
    url: http://catiny.com/yuvytung
  license:
    name: Unlicense
    url: http://www.apache.org/licenses/LICENSE-2.0.html
externalDocs:
  description: Find out more about Minh-Shop
  url: http://catiny.com
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://localhost:8080/api
    description: Development server with TLS Profile
  - url: http://catiny.com:8080/api
    description: Development server
  - url: https://catiny.com/api
    description: Product server with TLS Profile
#security:
#  - jwt: [ ]
#  - basic: [ ]
# TAGS ############################################################################################

tags:
  - name: video
    description: Everything about your videos
    externalDocs:
      description: Find out more
      url: http://catiny.com/open/api/file/videos
  - name: file
    description: Everything about your file
    externalDocs:
      description: Find out more the File
      url: http://catiny.com/open/api/files
# PATHS ###########################################################################################
paths:
  /open/file/videos/{videoName}:
    get:
      tags:
        - video
      summary: get file (video download data) by videoName
      description: api use to get data of a video by name of the video
      operationId: getVideoDataByName
      parameters:
        - name: videoName
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: video data
          content:
            video/mp4:
              schema:
                type: string
                format: byte
        404:
          description: File not found
          content: { }
  /open/file/videos/stream/{videoName}:
    get:
      tags:
        - video
      summary: get video ( stream data ) by videoName
      description: api use to get data (stream) of a video by name of the video
      operationId: getVideoStreamDataByName
      parameters:
        - name: videoName
          in: path
          required: true
          schema:
            type: string
        - name: Range
          in: header
          required: false
          schema:
            type: string
          description: this
      responses:
        200:
          description: video data
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        404:
          description: File not found
          content:
            application/problem+json:
              schema:
                $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'

  /open/files/{fileName}:
    get:
      tags:
        - file
      summary: get file (data) by fileName
      description: api use to get data of file by fileName of the File
      operationId: getFileDataByName
      parameters:
        - name: fileName
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: file data
          content:
            application/*: { }
        400:
          description: File not found
          content:
            application/problem+json:
              schema:
                $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'
  #############################################################################################
  /file/videos/upload:
    post:
      tags:
        - video
      summary: uploads an video or list video
      operationId: uploadVideos
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                videoDatas:
                  type: array
                  description: video to upload
                  items:
                    type: string
                    format: binary
            encoding:
              videoData:
                contentType: video/mp4
      responses:
        200:
          description: successfully uploaded
          content:
            application/json:
              schema:
                properties:
                  result:
                    type: object
        405:
          description: Invalid videoData value
          content: { }
      security:
        - jwt: [ ]
  /files/upload:
    post:
      tags:
        - file
      summary: uploads an file or list file
      operationId: uploadFiles
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                fileData:
                  type: array
                  description: data of file to upload
                  items:
                    type: string
                    format: binary
            encoding:
              videoData:
                contentType: video/mp4
      responses:
        200:
          description: successfully uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDefaultModel'
        405:
          description: Invalid fileData value
          content: { }
      security:
        - jwt: [ ]
components:
  schemas:
    ResponseDefaultModel:
      description: response default
      type: object
      properties:
        title:
          type: string
          default: Created successfully
          example: Created successfully
        type:
          type: string
          default: Create
          example: Create
        status:
          type: integer
          format: int32
          default: 200
          example: 200
        message:
          type: string
          default: Done.
        detail:
          type: string
          example: You have successfully uploaded
        path:
          type: string
          example: /api/file/upload
        id:
          type: integer
          format: int64
          example: 10194211
        uuid:
          type: string
          format: uuid
          example: 31c202af-5ea5-4e3c-9a3d-b00ba866ba25
        time:
          type: string
          format: date-time
          example: 2020-02-01 19:31:34.489910
        dataResponse:
          type: object
          description: đây là dữ liệu trả về kèm theo nếu có
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://petstore.swagger.io/oauth/dialog
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    jwt:
      name: jwt-token
      type: http
      description: JWT Authentication
      scheme: bearer
      bearerFormat: JWT
      in: header
    basic:
      type: http
      description: Basic Authentication
      scheme: basic
  responses:
    Problem:
      description: error occurred - see status code and problem object for more information.
      content:
        'application/problem+json':
          schema:
            $ref: 'https://opensource.zalando.com/problem/schema.yaml#/Problem'